pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pseudoflow-operator"
version = "0.3.0"
description = "Kubernetes operator for PseudoFlow CRD (structured layout, full DSL)"
requires-python = ">=3.11"
dependencies = [
  "kopf>=1.37.4",
  "kubernetes>=30.1.0",
  "pyyaml>=6.0.2",
  "jsonpath-ng>=1.6.1",
  "requests>=2.32.3",
]

[project.scripts]
pseudoflow-operator = "cmd.operator.cli:main"
